Adding files back from accidental deleting.
Made additional scripts for sorting abet files.
The v2 abet sorting checks for multiple recordings in addition to multiple abet files to sort.

Next step is to make the script look at and count number of avi files to better sort for multiple days/files
This can be done by counting and comparing number of frames.
Estimated by counting number of videos * 1000
OR last value of the timestamp file for frames, and comparing it to last value of ABET.

Updated minitouchsync script.
Created ParallelMiniTouchSync script that looks through for abet files and